engine/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compass-ai"
version = "0.1.0"
description = "Cursor for Product Managers — AI-native product discovery"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Compass AI"}]
keywords = ["product-management", "ai", "product-discovery", "cli"]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "chromadb>=0.4.0",
    "anthropic>=0.39.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
compass = "compass.cli:app"
compass-engine = "compass.server:main"

[tool.setuptools.packages.find]
include = ["compass*"]

[tool.ruff]
target-version = "py311"
line-length = 100
